Refactored and moved old MonthlySpending file into pages/Dashboard.jsx
