+++ src/pages/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./StylesCheatsheet.module.css";
+import styles from "./Styles.module.css";
 
 function MetricCard({ label, amount }) {
     return (
